добавлена фолбэк механика, в DTO добавлена дата создания json, специфичные DTO перенесены в сервис invest
