+++ invest_service/src/main/java/com/override/invest_service/dto/IMOEXDataDTO/IMOEXDataDTO.java
@@ -1,0 +1,17 @@
+package com.override.invest_service.dto.IMOEXDataDTO;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.*;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class IMOEXDataDTO {
+    private Analytics analytics;
+
+    @JsonProperty("analytics.dates")
+    private AnalyticsDates analyticsDates;
+}
