+++ invest_service/src/main/resources/moexReserveData/imoex.json
@@ -1,0 +1,86 @@
+{
+"analytics": {
+	"metadata": {
+		"indexid": {"type": "string", "bytes": 36, "max_size": 0},
+		"tradedate": {"type": "date", "bytes": 10, "max_size": 0},
+		"ticker": {"type": "string", "bytes": 36, "max_size": 0},
+		"shortnames": {"type": "string", "bytes": 189, "max_size": 0},
+		"secids": {"type": "string", "bytes": 36, "max_size": 0},
+		"weight": {"type": "double"},
+		"tradingsession": {"type": "int32"}
+	},
+	"columns": ["indexid", "tradedate", "ticker", "shortnames", "secids", "weight", "tradingsession"], 
+	"data": [
+		["IMOEX", "2024-04-12", "AFKS", "Система ао", "AFKS", 0.74, 3],
+		["IMOEX", "2024-04-12", "AFLT", "Аэрофлот", "AFLT", 0.66, 3],
+		["IMOEX", "2024-04-12", "AGRO", "AGRO-гдр", "AGRO", 0.8, 3],
+		["IMOEX", "2024-04-12", "ALRS", "АЛРОСА ао", "ALRS", 1.27, 3],
+		["IMOEX", "2024-04-12", "BSPB", "БСП ао", "BSPB", 0.37, 3],
+		["IMOEX", "2024-04-12", "CBOM", "МКБ ао", "CBOM", 0.66, 3],
+		["IMOEX", "2024-04-12", "CHMF", "СевСт-ао", "CHMF", 2.32, 3],
+		["IMOEX", "2024-04-12", "ENPG", "ЭН+ГРУП ао", "ENPG", 0.46, 3],
+		["IMOEX", "2024-04-12", "FEES", "Россети", "FEES", 0.45, 3],
+		["IMOEX", "2024-04-12", "FIVE", "FIVE-гдр", "FIVE", 1.02, 3],
+		["IMOEX", "2024-04-12", "FLOT", "Совкомфлот", "FLOT", 0.58, 3],
+		["IMOEX", "2024-04-12", "GAZP", "ГАЗПРОМ ао", "GAZP", 11.17, 3],
+		["IMOEX", "2024-04-12", "GLTR", "GLTR-гдр", "GLTR", 0.26, 3],
+		["IMOEX", "2024-04-12", "GMKN", "ГМКНорНик", "GMKN", 6.71, 3],
+		["IMOEX", "2024-04-12", "HYDR", "РусГидро", "HYDR", 0.24, 3],
+		["IMOEX", "2024-04-12", "IRAO", "ИнтерРАОао", "IRAO", 1.53, 3],
+		["IMOEX", "2024-04-12", "LKOH", "ЛУКОЙЛ", "LKOH", 15, 3],
+		["IMOEX", "2024-04-12", "MAGN", "ММК", "MAGN", 1.27, 3],
+		["IMOEX", "2024-04-12", "MGNT", "Магнит ао", "MGNT", 1.48, 3],
+		["IMOEX", "2024-04-12", "MOEX", "МосБиржа", "MOEX", 1.12, 3],
+		["IMOEX", "2024-04-12", "MSNG", "+МосЭнерго", "MSNG", 0.36, 3],
+		["IMOEX", "2024-04-12", "MTLR", "Мечел ао", "MTLR", 0.49, 3],
+		["IMOEX", "2024-04-12", "MTLRP", "Мечел ап", "MTLRP", 0.33, 3],
+		["IMOEX", "2024-04-12", "MTSS", "МТС-ао", "MTSS", 1.23, 3],
+		["IMOEX", "2024-04-12", "NLMK", "НЛМК ао", "NLMK", 1.89, 3],
+		["IMOEX", "2024-04-12", "NVTK", "Новатэк ао", "NVTK", 4.02, 3],
+		["IMOEX", "2024-04-12", "OZON", "OZON-адр", "OZON", 1.05, 3],
+		["IMOEX", "2024-04-12", "PHOR", "ФосАгро ао", "PHOR", 0.75, 3],
+		["IMOEX", "2024-04-12", "PIKK", "ПИК ао", "PIKK", 1.8, 3],
+		["IMOEX", "2024-04-12", "PLZL", "Полюс", "PLZL", 2.89, 3],
+		["IMOEX", "2024-04-12", "POSI", "iПозитив", "POSI", 0.4, 3],
+		["IMOEX", "2024-04-12", "ROSN", "Роснефть", "ROSN", 2.2, 3],
+		["IMOEX", "2024-04-12", "RTKM", "Ростел -ао", "RTKM", 1.06, 3],
+		["IMOEX", "2024-04-12", "RUAL", "РУСАЛ ао", "RUAL", 1.33, 3],
+		["IMOEX", "2024-04-12", "SBER", "Сбербанк", "SBER", 12.44, 3],
+		["IMOEX", "2024-04-12", "SBERP", "Сбербанк-п", "SBERP", 2.4, 3],
+		["IMOEX", "2024-04-12", "SELG", "Селигдар", "SELG", 0.35, 3],
+		["IMOEX", "2024-04-12", "SGZH", "Сегежа", "SGZH", 0.22, 3],
+		["IMOEX", "2024-04-12", "SMLT", "Самолет ао", "SMLT", 0.33, 3],
+		["IMOEX", "2024-04-12", "SNGS", "Сургнфгз", "SNGS", 3.24, 3],
+		["IMOEX", "2024-04-12", "SNGSP", "Сургнфгз-п", "SNGSP", 3.13, 3],
+		["IMOEX", "2024-04-12", "TATN", "Татнфт 3ао", "TATN", 5.7, 3],
+		["IMOEX", "2024-04-12", "TATNP", "Татнфт 3ап", "TATNP", 1.06, 3],
+		["IMOEX", "2024-04-12", "TCSG", "ТКСХолд ао", "TCSG", 0.59, 3],
+		["IMOEX", "2024-04-12", "TRNFP", "Транснф ап", "TRNFP", 0.61, 3],
+		["IMOEX", "2024-04-12", "UPRO", "Юнипро ао", "UPRO", 0.27, 3],
+		["IMOEX", "2024-04-12", "VKCO", "МКПАО \"ВК\"", "VKCO", 0.85, 3],
+		["IMOEX", "2024-04-12", "VTBR", "ВТБ ао", "VTBR", 0.9, 3]
+	]
+},
+"analytics.cursor": {
+	"metadata": {
+		"INDEX": {"type": "int64"},
+		"TOTAL": {"type": "int64"},
+		"PAGESIZE": {"type": "int64"},
+		"PREV_DATE": {"type": "string", "bytes": 30, "max_size": 0},
+		"NEXT_DATE": {"type": "undefined", "bytes": 0, "max_size": 0}
+	},
+	"columns": ["INDEX", "TOTAL", "PAGESIZE", "PREV_DATE", "NEXT_DATE"], 
+	"data": [
+		[0, 48, 100, "2024-04-11", null]
+	]
+},
+"analytics.dates": {
+	"metadata": {
+		"from": {"type": "date", "bytes": 30, "max_size": 0},
+		"till": {"type": "date", "bytes": 30, "max_size": 0}
+	},
+	"columns": ["from", "till"], 
+	"data": [
+		["2001-01-03", "2024-04-11"]
+	]
+}}
